guard line of sight + more cameras
