+++ Assets/Materials/Flashlight.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ff13497af3316a94498f81185ee04091
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
