+++ Assets/Prefabs/Enemies/PatrollingGuard.prefab
@@ -119,9 +119,7 @@ GameObject:
   - component: {fileID: 7909075326472278700}
   - component: {fileID: 2261609528579273078}
   - component: {fileID: 4705605659090736837}
-  - component: {fileID: 9057069766262060411}
-  - component: {fileID: 7297641987852475166}
-  - component: {fileID: 339775588482210895}
+  - component: {fileID: 834473241140816224}
   m_Layer: 0
   m_Name: PatrollingGuard
   m_TagString: Untagged
@@ -223,7 +221,7 @@ CapsuleCollider:
   m_Height: 2
   m_Direction: 1
   m_Center: {x: 0, y: 0, z: 0}
---- !u!114 &9057069766262060411
+--- !u!114 &834473241140816224
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -232,62 +230,12 @@ MonoBehaviour:
   m_GameObject: {fileID: 6501310672267175682}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: d5a57f767e5e46a458fc5d3c628d0cbb, type: 3}
+  m_Script: {fileID: 11500000, guid: 769ae04e0188f904d9ebefa614e005d8, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  GlobalObjectIdHash: 951099334
-  AlwaysReplicateAsRoot: 0
-  SynchronizeTransform: 1
-  ActiveSceneSynchronization: 0
-  SceneMigrationSynchronization: 1
-  SpawnWithObservers: 1
-  DontDestroyWithOwner: 0
-  AutoObjectParentSync: 1
---- !u!114 &7297641987852475166
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6501310672267175682}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: df2868252ab5c4d1da357e8f11f1b524, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  SyncPositionX: 1
-  SyncPositionY: 1
-  SyncPositionZ: 1
-  SyncRotAngleX: 0
-  SyncRotAngleY: 1
-  SyncRotAngleZ: 0
-  SyncScaleX: 0
-  SyncScaleY: 0
-  SyncScaleZ: 0
-  PositionThreshold: 0.001
-  RotAngleThreshold: 0.01
-  ScaleThreshold: 0.01
-  UseQuaternionSynchronization: 0
-  UseQuaternionCompression: 0
-  UseHalfFloatPrecision: 0
-  InLocalSpace: 0
-  Interpolate: 1
-  SlerpPosition: 0
---- !u!114 &339775588482210895
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6501310672267175682}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: efb258132035ff8458d8481a5dec7169, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  waypointsObject: {fileID: 0}
   moveSpeed: 5
   distanceThreshold: 0.1
+  alertTimer: 2
 --- !u!1 &6800124188568921770
 GameObject:
   m_ObjectHideFlags: 0
@@ -726,7 +674,9 @@ GameObject:
   m_Component:
   - component: {fileID: 3916278122732674762}
   - component: {fileID: 6087035560382154288}
-  - component: {fileID: 4060569273558692918}
+  - component: {fileID: 481360941308139354}
+  - component: {fileID: 6677530747828844111}
+  - component: {fileID: 1425389517653758295}
   m_Layer: 0
   m_Name: LineOfSight
   m_TagString: Untagged
@@ -743,8 +693,8 @@ Transform:
   m_GameObject: {fileID: 8605586554046433180}
   serializedVersion: 2
   m_LocalRotation: {x: 0.0000020191073, y: 0.7071068, z: 0.7071068, w: 0.0000020191073}
-  m_LocalPosition: {x: 0.00000012156, y: 0.41200024, z: 1.5}
-  m_LocalScale: {x: 0.925, y: 1.675, z: 1}
+  m_LocalPosition: {x: 0.00000012156, y: 0.41200024, z: 2.19}
+  m_LocalScale: {x: 0.925, y: 2.37984, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 3348841602411470165}
@@ -764,7 +714,7 @@ CapsuleCollider:
     serializedVersion: 2
     m_Bits: 0
   m_LayerOverridePriority: 0
-  m_IsTrigger: 0
+  m_IsTrigger: 1
   m_ProvidesContacts: 0
   m_Enabled: 1
   serializedVersion: 2
@@ -772,7 +722,7 @@ CapsuleCollider:
   m_Height: 2.0548952
   m_Direction: 1
   m_Center: {x: 0.00000006443745, y: 0.027445963, z: -0.00000011920929}
---- !u!114 &4060569273558692918
+--- !u!114 &481360941308139354
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -781,6 +731,56 @@ MonoBehaviour:
   m_GameObject: {fileID: 8605586554046433180}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 96197ab83c0ad7f4391781ab6dbbf5fd, type: 3}
+  m_Script: {fileID: 11500000, guid: 68b7dcb7e7cb204498fab1086dedd03c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!33 &6677530747828844111
+MeshFilter:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8605586554046433180}
+  m_Mesh: {fileID: 10206, guid: 0000000000000000e000000000000000, type: 0}
+--- !u!23 &1425389517653758295
+MeshRenderer:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8605586554046433180}
+  m_Enabled: 1
+  m_CastShadows: 1
+  m_ReceiveShadows: 1
+  m_DynamicOccludee: 1
+  m_StaticShadowCaster: 0
+  m_MotionVectors: 1
+  m_LightProbeUsage: 1
+  m_ReflectionProbeUsage: 1
+  m_RayTracingMode: 2
+  m_RayTraceProcedural: 0
+  m_RenderingLayerMask: 1
+  m_RendererPriority: 0
+  m_Materials:
+  - {fileID: 2100000, guid: ff13497af3316a94498f81185ee04091, type: 2}
+  m_StaticBatchInfo:
+    firstSubMesh: 0
+    subMeshCount: 0
+  m_StaticBatchRoot: {fileID: 0}
+  m_ProbeAnchor: {fileID: 0}
+  m_LightProbeVolumeOverride: {fileID: 0}
+  m_ScaleInLightmap: 1
+  m_ReceiveGI: 1
+  m_PreserveUVs: 0
+  m_IgnoreNormalsForChartDetection: 0
+  m_ImportantGI: 0
+  m_StitchLightmapSeams: 1
+  m_SelectedEditorRenderState: 3
+  m_MinimumChartSize: 4
+  m_AutoUVMaxDistance: 0.5
+  m_AutoUVMaxAngle: 89
+  m_LightmapParameters: {fileID: 0}
+  m_SortingLayerID: 0
+  m_SortingLayer: 0
+  m_SortingOrder: 0
+  m_AdditionalVertexStreams: {fileID: 0}
