+++ Assets/Prefabs/UI/AlertMeter.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e6e69a227d581034ab7d74cb03a83ab5
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
