--- Assets/RaycastForPlayer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 96197ab83c0ad7f4391781ab6dbbf5fd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
