+++ Assets/Scripts/Enemies/GuardLineOfSight.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 68b7dcb7e7cb204498fab1086dedd03c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
