+++ Assets/Scripts/Enemies/PatrollingGuard.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 769ae04e0188f904d9ebefa614e005d8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
