--- Assets/Scripts/Enemies/WaypointMover.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: efb258132035ff8458d8481a5dec7169
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
