+++ Assets/Scripts/Enemies/WaypointParent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: db55069305de4c3419794e093a9e50b2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
