+++ Assets/Sprites.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2ffc09bf4f3938b49a81b25a287a876e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
